fix: guard against missing missing status/gadget methods in ConferenceData
When ingesting calendar invites that include a videoconference link (i.e. via `conferenceData` in Google Calendar API), the gem was crashing with:

  undefined method `status` for Google::Apis::CalendarV3::ConferenceData

This is because not all `ConferenceData` objects implement the `status` or `gadget` methods, causing a NoMethodError during `jekyll build`.

This patch:
- Adds `respond_to?(:status)` guard before accessing `data.status`
- Adds `respond_to?(:gadget)` guard before accessing `data.gadget`

Thus, the generator will no longer crash on invites that include video links but omit those methods.

References:
- Google Calendar API `conferenceData` spec (events resource)  [oai_citation:0‡Google for Developers](https://developers.google.com/workspace/calendar/api/v3/reference/events)
- ConferenceData model documentation  [oai_citation:1‡Google for Developers](https://developers.google.com/resources/api-libraries/documentation/calendar/v3/java/latest/com/google/api/services/calendar/model/ConferenceData.html)